+++ CHANGELOG.md
@@ -1,1 +1,9 @@
-Initial release
+[View Full Changelog](https://github.com/enderneko/BFInfinite/compare/r1-alpha...5cf96dcc2bb475abe1130d863bc4b40c824cd456)
+
+- Add class accent color support
+- Add AF version check on load
+- Fix UF preset issue
+- Temporary font fix
+- Update buffsDebuffs options
+- Update font options
+- Use complement color for logo gradient